+++ cli/abeamer-cli.ts
@@ -82,6 +82,7 @@ import { HttpServerEx } from "../shared/vendor/http-server-ex.js";
  *
  *  Creates a movie from the previous generated image sequence on `foo/story-frames/movie.mp4`.
  * `abeamer movie foo/`.
+ *  Requires that `ffmpeg` to be on the search path, or set `FFMPEG_BIN=<absolute-path-to-executable>`
  *
  *  Creates the movie `foo/story.webm`.
  * `abeamer movie foo/ --movie foo/story.webm`.
@@ -600,7 +601,8 @@ To modify the fps, edit the [js/main.ts] file.
     const movieFileName = argOpts['movie'].value as string
       || `${report.dirname}/${DEFAULT_MOVIE_NAME}`;
     const bkgMovieFileName = argOpts['bkgMovie'].value as string;
-    const cmdLine = 'ffmpeg';
+    const cmdLine = sysProcess.env['FFMPEG_BIN'] || 'ffmpeg';
+
     let args = ['-r', report.fps.toString(), '-f', 'image2',
       '-s', `${report.width}x${report.height}`,
       '-i', report.framespattern,