+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   > These arguments to be passed to convert before/after the arguments passed by abeamer
 - (DEV) Add `Launch abeamer cli` to VSC debugger configuration.
 - (DEV) Add excludes to vsc files.watcher
+- Add support of system env FFMPEG_BIN to locate ffmpeg executable.
 
 ### Fix
 - Fix http-server for URL with path with %20(spaces).